Move Home from components to pages folder
